Move event loop to own file for easy lookup
